Updates to organization schema and interface
