+++ packages/types/src/organization.ts
@@ -5,9 +5,24 @@ import { z } from 'zod';
 
 /* * */
 
+export const OrganizationValueSchema = z.union([
+	z.string(),
+	z.string(),
+	z.boolean(),
+	z.array(z.string()),
+	z.array(z.string()),
+]);
+
+export type OrganizationValue = z.infer<typeof OrganizationValueSchema>;
+
+/* * */
+
 export const OrganizationSchema = DocumentSchema.extend({
-	code: z.string(),
-	name: z.string(),
+	home_links: z.array(z.string()).default([]),
+	home_wikis: z.array(z.string()).default([]),
+	logo: z.string().nullish(),
+	theme: z.string().nullish(),
+
 }).strict();
 
 export const CreateOrganizationSchema = OrganizationSchema.omit({ _id: true, created_at: true, updated_at: true });
